Rename server file to match class name
